Basic Gameplay Complete
Added invisible walls to the scene such that if the player moves too much to the left or right he does not fall over the plane or into the gutter. Added player layers to the player object and invisible walls, and added ball layers to the ball object so that if the ball hits the walls it ignores it and still interacts with the gutter.